# docs/swagger.yaml
basePath: /api/v1
definitions:
  approval-engine-service_internal_domain.Application:
    properties:
      api_key:
        type: string
      callback_url:
        type: string
      code:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  approval-engine-service_internal_domain.Participant:
    properties:
      department:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      position:
        type: string
      superior_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  approval-engine-service_pkg_response.Envelope:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  approval-engine-service_pkg_response.Page:
    properties:
      items: {}
      meta:
        $ref: '#/definitions/approval-engine-service_pkg_response.PageMeta'
    type: object
  approval-engine-service_pkg_response.PageMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_handler.CreateApplicationBody:
    properties:
      callback_url:
        example: https://assetmgmt.internal/api/approval-webhooks
        type: string
      code:
        example: assetmgmt
        type: string
      name:
        example: Asset Management
        type: string
    type: object
...
  internal_handler.ImportBody:
    properties:
      participants:
        items:
          $ref: '#/definitions/approval-engine-service_internal_domain.Participant'
        type: array
    type: object
...
info:
  contact: {}
  description: |-
    Reusable approval workflow engine shared across Mayora internal systems (asset management first). A consuming app's requester_id/user_id are always checked against registered, active participants — the engine never trusts an identity blindly.
    Auth: every /requests route requires X-API-Key (a consuming application's server-to-server key — never expose it in a browser), and an application only ever sees and decides its own requests. /participants/import requires the operator X-Admin-Key. /portal/* serves the browser portal and can be switched off with PORTAL_ENABLED=false until SSO login is in place.
  title: Approval Engine API
  version: "1.0"
paths:
...
  /participants/import:
    post:
      consumes:
      - application/json
      description: 'Upserts by user_id (NIK) in one transaction — a partial failure
        never leaves the table half-updated. Rows may reference a superior_id that
        appears later in the same array (or not at all yet); foreign keys are deferred
        to commit time, so the array does not need to be topologically sorted. Operator-only
        (X-Admin-Key): the org chart is shared by every consuming application, so
        no single application''s X-API-Key may rewrite it.'
      parameters:
      - description: Participants to import
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ImportBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "401":
          description: missing or invalid X-Admin-Key
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "503":
          description: ADMIN_API_KEY not configured on this deployment
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      security:
      - AdminKeyAuth: []
      summary: Import participants (e.g. from a FICOM export)
      tags:
      - participants
  /portal/applications:
    get:
      description: Never includes api_key — this is a read view for the management
        UI, not a way to recover a lost key.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_pkg_response.Page'
              type: object
      summary: List consuming applications
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: 'Generates a new X-API-Key for the application. The key is only
        ever returned in this one response — List never includes it, so it must be
        copied to the consuming app''s config immediately. callback_url is optional:
        if set, the engine POSTs a signed webhook there on every request/step state
        change (see WebhookEvent); if left empty the app can still always poll GET
        /requests/{id}. No role gate yet (accepted gap for the demo scope).'
      parameters:
      - description: Application to register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateApplicationBody'
      produces:
      - application/json
      responses:
        "201":
          description: includes api_key, shown once
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.Application'
              type: object
        "400":
          description: validation error or duplicate code
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      summary: Register a consuming application
      tags:
      - applications
  /portal/inbox/{userID}:
    get:
      description: Everything pending this approver's decision, across every consuming
        application — the "one inbox for every system" promise of a centralized approval
        portal.
      parameters:
      - description: Approver NIK
        in: path
        name: userID
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_pkg_response.Page'
              type: object
      summary: Approver inbox
      tags:
      - inbox
  /portal/participants:
    get:
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_pkg_response.Page'
              type: object
      summary: List participants
      tags:
      - participants
...
securityDefinitions:
  AdminKeyAuth:
    in: header
    name: X-Admin-Key
    type: apiKey
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"